backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-text-splitters
chromadb
requests
pypdf
python-multipart
ollama
pytesseract
pdf2image
pillow
python-docx


